move translation selection dialog to a new component
